Module legacy_consensus::blockmania::block_minter

Structs

BlockBuffer
BlockBuffer is the EventConsumer to BlockBuilder's minter This stores events in a time-buffered fashion and deterministically creates blocks for nodes to compute on.

BlockBuilder
The main storage structure for block creation and storage

BlockHash

LatestHashProposal
This instance of BlockProposal fixes the struct used to represent Blocks created from GraphEvents.

RecordStore
We structure the storage of input records as a vector with pointers to the index of this vector. This is notionally equivalent to a HashMap<HashValue, VerifiedEvent<Payload>>, but it allows us to track the ingestion order of those events. Since we take assumptions that these are in one specific topological order w.r.t. chaining, it allows us to manipulate (e.g. output) these events while respecting the topological order

Traits

EventConsumer