Module legacy_consensus::blockmania::block_proposal

Structs

BlockRef

Position
In BlockMania, blocks are proposed for a specific Index, relative to a certain Author.

ProposalId
A more expressive type for hash values, hinting at which position a hash is suggested for. Since the HashValue here depends on the author and index, note that specifying Eq and Hash for the trait object let us query ProposalId-indexed structures with a HashValue

ScopedPosition

Traits

BlockProposal
The central trait for what a block proposal looks like in BlockMania. This is a close analogous to a VerifiedEvent.

HashKeyed
A technical trait that serves both:

Type Definitions

View
A View is just a counter