Trait legacy_consensus::blockmania::block_proposal::BlockProposal

pub trait BlockProposal {
    fn hash(&self) -> &HashValue;

    fn author(&self) -> &Author;

    fn round(&self) -> &Index;

    fn references(&self) -> Vec<BlockRef>;

    fn make_id(&self) -> ProposalId { ... }

    fn max_height(&self) -> Index { ... }
}

The central trait for what a block proposal looks like in BlockMania. This is a close analogous to a VerifiedEvent.

Required methods

fn hash(&self) -> &HashValue

Block proposals are indexed by the block hash.

fn author(&self) -> &Author

Block proposals need to be positioned.

fn round(&self) -> &Index

fn references(&self) -> Vec<BlockRef>

Block proposals refer to other blocks.

Loading content...
Provided methods

fn make_id(&self) -> ProposalId

We provide a way to get the corresponding ProposalId

fn max_height(&self) -> Index

We can compute the max height of this proposal, which is the largest round over the transitive closure of references.

It is an explicit purpose of this not to compute the transitive closure at each call, but to rather rely on the max_height indexed in references, as this allows purging memory of committed proposals.

Loading content...