Trait legacy_consensus::blockmania::block_proposal::HashKeyed

pub trait HashKeyed {
    fn hash_key(&self) -> &HashValue;
}

A technical trait that serves both:

as a type family describing the range of queries against a BlockStore,

as a way to make explicit how the Hash, Eq of a ProposalId is constructed.

In concrete terms, this lets us express everything in store in terms of a ProposalId, while only paying the costs of a HashValue.

Required methods

fn hash_key(&self) -> &HashValue

Loading content...