Module legacy_consensus::blockmania::block_state

Structs

BlockBook

BlockState
This struct outlines the view of the world as seen from the point of view of a single position, hereafter "the local point of view". See below for the global view, stored in the object State. It allows the state machines of block interpretation to only move forward in time, and to be cleaned up globally by timeout. Their snapshots also allow for a trace / log of block interpretation at a very granular level. They are cleaned once the interpretation of blocks is final (see the Positioned trait and the cleanup procedure below).

State
This struct implements state annotation for Blockmania: from BlockDAGs to Consensus, by G. Danezis & D. Hrycyszyn (2018) 1.

Status

Enums

VirtualMessage
This enum outlines the type of virtual messages generated by nodes along the annotation process. They are not actually exchanged over the network.

Traits

Positioned