Struct legacy_consensus::blockmania::parent_pointer_tree::ParentPointerTree

pub struct ParentPointerTree<T> { /* fields omitted */ }

An immutable parent pointer node. May be empty or contain a value; may have a pointer to a parent or not.

A parent pointer tree is a (possibly empty) node with a (possibly null) pointer to a parent node. Any given node has a unique path back to the root node. Rather than mutably updating the stack, one creates and obtains access to immutable nodes (when a node becomes unreachable its memory is automatically reclaimed). A new child node pointing to a parent can be created via the child function (analogous to the normal push) and a parent can be retrieved via the parent function (analogous to the normal pop).

Methods

impl<T> ParentPointerTree<T>

pub fn new() -> ParentPointerTree<T>

Return an empty parent pointer tree node.

pub fn is_empty(&self) -> bool

Is this parent pointer tree node empty?

pub fn len(&self) -> usize

How many items are there in this parent pointer tree?

pub fn child(&self, val: T) -> ParentPointerTree<T>

Create a new parent pointer tree node containing value val and pointing to parent self.

pub fn parent(&self) -> Option<ParentPointerTree<T>>

Return this parent pointer tree node's parent node or None if this parent pointer tree is empty.

pub fn val(&self) -> Option<&T>

Return a reference to this parent pointer tree node's value or None if this parent pointer tree is empty.

Important traits for ParentPointerTreeValsIter<'a, T>
impl<'a, T> Iterator for ParentPointerTreeValsIter<'a, T> type Item = &'a T;

pub fn vals(&self) -> ParentPointerTreeValsIter<T>

Return an iterator over this parent pointer tree's values. Note that the iterator produces values starting from this node and then walking up towards the root.

pub fn try_unwrap(self) -> Result<T, ParentPointerTree<T>>

Try to consume this ParentPointerTree node and return its data. If the parent pointer tree node has no children, this succeeds; if the parent pointer tree node has children, it fails, and returns the original parent pointer tree node.