Trait legacy_consensus::graph_consensus::events::ConsensusEventProcessor

pub trait ConsensusEventProcessor {
type Payload: Serialize;
    fn reset(&mut self, pool_size: usize);

    fn process_verified_event(&mut self, event: VerifiedEvent<Self::Payload>);

    fn get_processed_event(
        &self, 
        event_hash: &HashValue
    ) -> Option<&VerifiedEvent<Self::Payload>>;

    fn next_committed_payload(&mut self) -> Option<Committed<Self::Payload>>;

    fn graphviz_output(
        &self, 
        my_name: &str, 
        author_to_name: &HashMap<Author, String>
    ) -> Result<String, Error>;

    fn matching_events(
        &self, 
        query: &EventQuery
    ) -> (Vec<GraphEvent<Self::Payload>>, Moniker);

    fn current_moniker(&self) -> Moniker;

    fn verify(
        &self, 
        event: GraphEvent<Self::Payload>, 
        initial_hash: &HashValue, 
        trusted_nodes: &HashMap<Author, PublicKey>
    ) -> EventStatus<Self::Payload> { ... }
}

A consensus engine that works by processing a DAG of events offline.

Associated Types

type Payload: Serialize

Loading content...
Required methods

fn reset(&mut self, pool_size: usize)

Reset (or set) the parameters of the consensus engine. All events not yet returned by next_committed_payload will be re-inserted and processed with the new consensus parameters.

Reset points and their parameters must be strictly agreed upon between participants. Therefore, in practice, this method will likely be triggered only at the beginning or when a special message was just returned by next_committed_payload.

fn process_verified_event(&mut self, event: VerifiedEvent<Self::Payload>)

Process a new event. We assume that all the necessary verifications have been done already.

fn get_processed_event(
&self,
event_hash: &HashValue
) -> Option<&VerifiedEvent<Self::Payload>>

Look up whether an event was already processed. This is used for event chaining verifications before insertion.

fn next_committed_payload(&mut self) -> Option<Committed<Self::Payload>>

Iterate and return on the next committed payload to process. This will trigger clean-ups.

fn graphviz_output(
&self,
my_name: &str,
author_to_name: &HashMap<Author, String>
) -> Result<String, Error>

Generate the Graphviz output for this event processor

fn matching_events(
&self,
query: &EventQuery
) -> (Vec<GraphEvent<Self::Payload>>, Moniker)

Process a DAG-synchronization query and return a moniker on the next data.

fn current_moniker(&self) -> Moniker

Get a description of the latest available data.

Loading content...
Provided methods

fn verify(
&self,
event: GraphEvent<Self::Payload>,
initial_hash: &HashValue,
trusted_nodes: &HashMap<Author, PublicKey>
) -> EventStatus<Self::Payload>

Verify that the graph event can be processed now.

Loading content...