Trait legacy_consensus::graph_consensus::ConsensusProvider

pub trait ConsensusProvider {
type Payload;
    fn start(
        &mut self, 
        commit_callback: UnboundedSender<Committed<Self::Payload>>
    ) -> Sender<Self::Payload>;

    fn stop(&mut self);

    fn graphviz_output(
        &self, 
        my_name: String, 
        author_to_name: HashMap<Author, String>
    ) -> Box<dyn Future<Item = String, Error = Error>>;
}

Public interface to a consensus protocol.

Associated Types

type Payload

Loading content...
Required methods

fn start(
&mut self,
commit_callback: UnboundedSender<Committed<Self::Payload>>
) -> Sender<Self::Payload>

Start a consensus thread using the given callback channel to send committed payloads. The call returns a channel to be used to submit inputs to the consensus protocol. We use a SyncSender to provide back-pressure in the future.

fn stop(&mut self)

Stop the consensus threads.

fn graphviz_output(
&self,
my_name: String,
author_to_name: HashMap<Author, String>
) -> Box<dyn Future<Item = String, Error = Error>>

Loading content...