Module legacy_consensus::state_replication

Modules

mocked_commit

mocked_state

replication_from_graph_consensus

Traits

CommandSender

QuorumVerifier
How to determine that a set of public keys form a quorum.

SecureCommit
How to build and verify a proof of finality for a particular state.

StateComputer

StateFinalizer
How to communicate that a state was committed or discarded.

StateMachineReplicationProvider
Public interface to a state-machine-replication protocol.

Functions

make_state_replication_provider
Create a state-replication provider based a "graph-consensus" protocol.