Trait legacy_consensus::state_replication::SecureCommit

pub trait SecureCommit {
type State;
    fn new(state: Self::State) -> Self;

    fn state(&self) -> &Self::State;

    fn sign(&mut self, public_key: &PublicKey, private_key: &PrivateKey);

    fn merge(&mut self, other: Self) -> Result<(), Error>;

    fn valid_signers<'a>(
        &'a self
    ) -> Box<dyn Iterator<Item = &'a PublicKey> + 'a>;
}

How to build and verify a proof of finality for a particular state.

Associated Types

type State

The state being signed.

Loading content...
Required methods

fn new(state: Self::State) -> Self

Create an initial commit for the given state.

fn state(&self) -> &Self::State

Get the state.

fn sign(&mut self, public_key: &PublicKey, private_key: &PrivateKey)

Operate the given private key to add a signature to the commit.

fn merge(&mut self, other: Self) -> Result<(), Error>

Import the signatures of 'other' into 'self'. The states must match.

fn valid_signers<'a>(&'a self) -> Box<dyn Iterator<Item = &'a PublicKey> + 'a>

Verify signatures and return the identities of valid signers, if any.

Loading content...