Trait legacy_consensus::state_replication::StateMachineReplicationProvider

pub trait StateMachineReplicationProvider {
type State: QuorumVerifier;
type Command;
type SecureCommit: SecureCommit;
    fn start(
        &mut self, 
        initial_state: Self::State, 
        compute_state_callback: Box<dyn StateComputer<Self::State, Self::Command>>, 
        finalize_state_callback: Box<dyn StateFinalizer<Self::State, Self::SecureCommit>>
    ) -> Box<dyn CommandSender<Self::Command>>;

    fn stop(&mut self);
}

Public interface to a state-machine-replication protocol.

Such a protocol is meant to encapsulate

shared mempool abstractions (if any)

transaction selection

Byzantine-Fault-Tolerant consensus protocols (leader-based or not)

checkpointing protocols (if any) to gather proofs of finality

Associated Types

type State: QuorumVerifier

Abstract State to replicate. In practice, this should be a reference to storage. For testing, it could be a map of values. It will likely include a timestamp and additional information needed for final verifications by end users. A state should also carry information about the trusted nodes (see QuorumVerifier trait)

type Command

One or several abstract transactions to execute on top of a state, thus creating a new one.

type SecureCommit: SecureCommit

Proof of finality for a state. The validation of a SecureCommit value returned below is guaranteed to succeed, meaning that a quorum of nodes have obtained the same state after having executed the same sequence of command, and have agreed that this sequence is final. To avoid inconsistencies in case of a crash, SecureCommit values are intended to be safely stored and indexed in local long term storage (aka "blockchain") immediately after being returned by consensus, preferably before making them visible to the end users of this node.

Loading content...
Required methods

fn start(
&mut self,
initial_state: Self::State,
compute_state_callback: Box<dyn StateComputer<Self::State, Self::Command>>,
finalize_state_callback: Box<dyn StateFinalizer<Self::State, Self::SecureCommit>>
) -> Box<dyn CommandSender<Self::Command>>

Start a protocol client using the given callbacks to manage commands, execution, and commit generation. Returns channel to send local commands.

fn stop(&mut self)

Stop the protocol client.

Loading content...